+++ usp-core-waap-header-filtering/step1/verify.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+# SPDX-FileCopyrightText: 2025 United Security Providers AG, Switzerland
+#
+# SPDX-License-Identifier: GPL-3.0-only
+
+test -e /tmp/.backend_installed